[RENAME] 디렉토리 명 오타 수정
